add better video activation
previously we only activated an article's video when it "entered" view,
meaning unless it was totally out of view and then became in view, we
wouldn't activate it's video — this prevents the "activation bar" effect
where you can scroll up/down over a horizontal point to toggle between
two articles' videos, and meant you could often be watching the wrong
video behind an article if you scrolled erratically.

this change adds the toggling behavior.

note that we can only toggle on each event in one direction, otherwise
the video activation function would be called as a new article entered
the view *and* the old article exited, which would be two calls to show
the same video.